+++ tests/testthat/test-git-commits.R
@@ -4,17 +4,21 @@ git_commits_branch <- str_c("test-git-commits-", format(Sys.time(), "%Y-%m-%d-%H
 master_sha <- view_shas(refs = "master", repo = "ChadGoymer/test-githapi")[[1]]
 
 setup(suppressMessages({
+
   create_branches(
     branches = git_commits_branch,
     shas     = master_sha,
     repo     = "ChadGoymer/test-githapi")
+
 }))
 
-teardown(suppressMessages({
+teardown(suppressMessages(tryCatch({
+
   delete_branches(
     branches = git_commits_branch,
     repo     = "ChadGoymer/test-githapi")
-}))
+
+})))
 
 
 # TEST: view_commits --------------------------------------------------------------------------